fix: Fix order

+++ src/main/java/com/greatminds/ayni/shopping/interfaces/rest/resources/CreateOrderResource.java
@@ -2,5 +2,5 @@ package com.greatminds.ayni.shopping.interfaces.rest.resources;
 
 import java.util.Date;
 
-public record CreateOrderResource(String description, Double totalPrice, Integer quantity, String paymentMethod, String status, Long saleId, Long orderedBy, Long acceptedBy, Date orderedDate) {
+public record CreateOrderResource(String description, Double totalPrice, Long quantity, String paymentMethod, String status, Long saleId, Long orderedBy, Long acceptedBy, Date orderedDate) {
 }
